pircbotx is getting an incorrect total number of users.
The output of the command in its current state can be seen here: http://i.jophest.us/dCX6.jpg
However, if we look to the user list: http://i.jophest.us/HgRV.jpg
The ops amount + total user amount = the amount pircbotx seems to be getting. To make sure of this I manually counted the amount of users to make sure it was not an error on hexchat's behalf.